+++ src/nplinker/genomics/__init__.py
@@ -4,17 +4,17 @@ import logging
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 
 
-from .BGC import BGC
-from .GCF import GCF
+from .bgc import BGC
+from .gcf import GCF
 from .genomics import loadBGC_from_cluster_files
 from .genomics import make_mibig_bgc_dict
-from .MiBIGBGC import MiBIGBGC
+from .mibigbgc import MiBIGBGC
 
 
 __all__ = [
-    "BGC",
-    "MiBIGBGC",
-    "GCF",
+    "bgc",
+    "mibigbgc",
+    "gcf",
     "loadBGC_from_cluster_files",
     "make_mibig_bgc_dict"
 ]
