+++ src/nplinker/genomics/bgc.py
