+++ src/nplinker/metabolomics/__init__.py
@@ -7,15 +7,15 @@ logging.getLogger(__name__).addHandler(logging.NullHandler())
 from .metabolomics import load_dataset
 from .metabolomics import load_edges
 from .metabolomics import mols_to_spectra
-from .MolecularFamily import MolecularFamily
-from .SingletonFamily import SingletonFamily
-from .Spectrum import Spectrum
+from .molecular_family import MolecularFamily
+from .singleton_family import SingletonFamily
+from .spectrum import Spectrum
 
 
 __all__ = [
-    "Spectrum",
-    "MolecularFamily",
-    "SingletonFamily",
+    "spectrum",
+    "molecular_family",
+    "singleton_family",
     "load_dataset",
     "load_edges",
     "mols_to_spectra"
